Support ghc 9.10 and prior (different support for exception annotations)
